drm/msm: enable seamless dynamic framerates
Add a DRM_MODE_FLAG to explicitly request seamless mode
changes. If this flag is set, enable/disable are not
called. Components check at mode_fixup if seamless mode is
possible, and what type of seamless transition. Dynamic
FPS control is the first feature to use this flag, which
allows seamlessly changing the FPS without a full modeset.

Change-Id: I87cf32a445f541f85cd3e83793fa847903e3d56d
Signed-off-by: Lloyd Atkinson <latkinso@codeaurora.org>